requirements.txt
# Core document processing
pymupdf>=1.24.0
Pillow>=10.0.0
pytesseract>=0.3.13
opencv-python-headless>=4.9.0
numpy>=1.24.0

# Barcode / QR detection
pyzbar>=0.1.9

# Data modeling & validation
pydantic>=2.0.0
python-dateutil>=2.8.2
PyYAML>=6.0

# Data processing & scoring
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.66.0

# Reporting
jinja2>=3.1.0

# UI
streamlit>=1.40.0
